piano-sheet-generator-from-youtube/requirements.txt
yt-dlp
opencv-python
numpy
img2pdf
Pillow
scikit-image
